Added new APIs in main-books branch

+++ src/main/java/com/example/employee/dto/BooksDTO.java
@@ -13,7 +13,7 @@ public class BooksDTO
     private String name;
     private String author;
     private String description;
-    private int numberOfPages;
+    private Long numberOfPages;
     private double price;
     private int bookCreatedBy;
     private boolean isDeleted;
